backend/requirements.txt
Django==2.2.16
requests==2.26.0
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
requests==2.26.0
PyJWT==2.4.0
asgiref==3.3.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
django-templated-mail==1.1.1
djoser==2.1.0
Pillow==9.0.0
reportlab==3.6.12
django-colorfield==0.7.2
webcolors==1.12
django-cors-headers==3.13.0